refactor: modularize core and supporting services layout
- update imports/paths
