+++ Core/app/schemas/auth.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+# Cхема проверяет, то ли прислал пользователь при регистрации или входе
+class UserRegister(BaseModel):
+    username: str
+    password: str
